PHPStan seviye 7: "false donebilir" her yerde cevaplandi
Seviye 7 birlesim tiplerini ciddiye aliyor: bir ifade X|false ise,
X bekleyen bir yere gecirilemez. 125 bulgu vardi. 58'i sorgu katmani
ile kendiliginden yok oldu (onceki commit); kalan 67'si burada.

UC SORU, UC FARKLI CEVAP

  1. "BU ALAN NE ZAMAN VAR?"  - 21 bulgu
     attach_store() ve mail_sender() sunu donduruyordu:
        array{ok: bool, error?: string, stored?: string, ...}
     Bu "hicbir alan garanti degil" demek. Oysa gercek su: ok=true
     ise dosya alanlari HER ZAMAN dolu, ok=false ise error HER ZAMAN
     dolu. Iki sekli ayirinca if ($r['ok']) kontrolu tipi de
     daraltiyor ve olmayan bir alana uzanmak yazarken yakalaniyor.

  2. "VARSAYILAN VARSA NULL DONER MI?"  - 4 bulgu
     input_enum('sort', array_keys($map), 'due') asla null donmuyor
     ama imza ?string diyordu. Dort ekran $map[$sort] yaziyor ve
     hakliydi - yazili olmayan bir varsayima guveniyorlardi.
     Kosullu donus tipi (@return ($default is null ? ... )) o
     varsayimi yaziya dokuyor.

  3. "GERCEKTEN BASARISIZ OLABILIR MI?"  - 27 bulgu
     KURAL: olabiliyorsa KORUMA, olamiyorsa CAST + neden olamadigi.
       fopen('php://output')  -> koruma. Basliklar gonderilmis
                                 durumda; yarim CSV uretmektense
                                 500 ile durmak durust.
       glob()                 -> ?: []   (okunamayan dizin mumkun)
       file_get_contents()    -> koruma  (izin sorunu mumkun)
       session_name() okuma   -> cast    (yalnizca ATAMA basarisiz olur)
       json_encode(dizi)      -> koruma  (gecersiz UTF-8 mumkun)
       strtotime("7 days")    -> cast    (sabit ifade)

GD'DE GERCEK BIR HATA IHTIMALI
  imagecolorallocatealpha() palet dolarsa false doner ve o deger
  dogrudan imagesetpixel()'e giriyordu - false sessizce "renk 0"
  diye yorumlanip ikonu bozardi. Tek bir color() yardimcisi kondu:
  false'ta duruyor, bilesenleri araliga kirpiyor (yuvarlamadan 256
  cikmasi gecerli bir renk degil), sekiz cagri yerini kapatiyor.

YAN KAZANIMLAR
  audit_value_short()  denetim ekranindaki ic ice uclu ifade bir
                       fonksiyona cikti - okunurluk ayrica kazandi
  flash_take()         oturumdan gelen bicimsiz girdiyi artik
                       eliyor: bozuk bir oturum dosyasi layout'u
                       yarim bir diziye uzandirmiyor
  discussion_*()       fetchAll() yerine db_all() - tipli
  SchemaTest::$pdo     sorgu katmanindan sonra olu alandi, kalkti

DOGRULAMA
  PHPStan seviye 7   0 hata          php-cs-fixer  0/142
  PHPUnit            159 + 33        kurallar      7/7
  smoke 59/59 · debug 109/109 · HTTP okuma 19/19 · HTTP yazma 9/9

Siradaki: seviye 8 (null guvenligi).

+++ tools/seed_demo.php
@@ -363,9 +363,9 @@ foreach ($actions as $j => [$titlePart, $actionTitle, $priority, $aStatus, $dueO
         ':owner'  => $userIds[$j % count($userIds)],
         ':pri'    => $priority,
         ':status' => $aStatus,
-        ':due'    => date('Y-m-d', strtotime("{$dueOffset} days")),
+        ':due'    => date('Y-m-d', (int)strtotime("{$dueOffset} days")),
         ':done'   => $aStatus === 'Completed'
-                       ? date('Y-m-d H:i:s', strtotime(($dueOffset - 2) . ' days'))
+                       ? date('Y-m-d H:i:s', (int)strtotime(($dueOffset - 2) . ' days'))
                        : null,
         ':cby'    => 1,
     ]);
